Correções 1

+++ src/main/java/model/Bank.java
@@ -68,7 +68,8 @@ public class Bank extends ReceiverAdapter {
                 TransferDTO transferDTO = (TransferDTO) messageDTO.getObject();
                 if (!transferDTO.getSource().getOwner().getId().equals(this.user.getId())) {
                     AccountController accountController = new AccountController();
-                    accountController.transferValue(transferDTO.getSource(), transferDTO.getTarget().getOwner().getCpf(), transferDTO.getValue());
+                    accountController.transferValue(transferDTO.getSource(),
+                            transferDTO.getTarget().getOwner().getCpf(), transferDTO.getValue());
                     userFile = new File(Constants.userPathFile);
                     accountFile = new File(Constants.accountPathFile);
                     try {
@@ -77,6 +78,7 @@ public class Bank extends ReceiverAdapter {
                         MessageDTO messageDTO1 = new MessageDTO();
                         messageDTO1.setExecutedClass(EnumExecutedClass.REPLICATE);
                         messageDTO1.setObject(new ArrayList<>(Arrays.asList(userContent, accountContent)));
+                        messageDTO1.setUserId(this.user.getId());
                         this.send(messageDTO1);
                     } catch (Exception e) {
                         e.printStackTrace();
@@ -84,22 +86,24 @@ public class Bank extends ReceiverAdapter {
                 }
                 break;
             case REPLICATE:
-                try {
-                    Files.deleteIfExists(new File(Constants.accountPathFile).toPath());
-                    Files.deleteIfExists(new File(Constants.userPathFile).toPath());
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                ArrayList<byte[]> contents = (ArrayList<byte[]>) messageDTO.getObject();
-                try (FileOutputStream fos = new FileOutputStream(Constants.userPathFile)) {
-                    fos.write(contents.get(0));
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-                try (FileOutputStream fos = new FileOutputStream(Constants.accountPathFile)) {
-                    fos.write(contents.get(1));
-                } catch (Exception e) {
-                    e.printStackTrace();
+                if (!this.user.getId().equals(messageDTO.getUserId())) {
+                    try {
+                        Files.deleteIfExists(new File(Constants.accountPathFile).toPath());
+                        Files.deleteIfExists(new File(Constants.userPathFile).toPath());
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                    ArrayList<byte[]> contents = (ArrayList<byte[]>) messageDTO.getObject();
+                    try (FileOutputStream fos = new FileOutputStream(Constants.userPathFile)) {
+                        fos.write(contents.get(0));
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                    try (FileOutputStream fos = new FileOutputStream(Constants.accountPathFile)) {
+                        fos.write(contents.get(1));
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
                 }
                 break;
         }
